Batch sumcheck verification moved from intmul to generic sumcheck.
