+++ crates/prover/src/protocols/sumcheck/batch.rs
@@ -5,11 +5,33 @@ use binius_transcript::{
 	ProverTranscript,
 	fiat_shamir::{CanSample, Challenger},
 };
-use binius_verifier::protocols::sumcheck::common::{BatchSumcheckOutput, RoundCoeffs};
+use binius_verifier::protocols::sumcheck::common::RoundCoeffs;
 
 use crate::protocols::sumcheck::{common::SumcheckProver, error::Error};
 
+/// Prover view of the execution result of a batched sumcheck.
+#[derive(Debug, PartialEq, Eq)]
+pub struct BatchSumcheckOutput<F: Field> {
+	/// Verifier challenges for each round of the sumcheck protocol.
+	///
+	/// One challenge is generated per variable in the multivariate polynomial,
+	/// with challenges\[i\] corresponding to the i-th round of the protocol.
+	///
+	/// Note: reverse when folding high-to-low to obtain evaluation claim.
+	pub challenges: Vec<F>,
+	/// Evaluation claims on non-transparent multilinears, per prover.
+	/// These values are concatenated and written to the transcript.
+	pub multilinear_evals: Vec<Vec<F>>,
+}
+
 /// Prove a batched sumcheck protocol execution, where all provers have the same number of rounds.
+///
+/// The batched sumcheck reduces a set of claims about the sums of a multivariate polynomials over
+/// the boolean hypercube to their evaluation at a (shared) challenge point. This is achieved by
+/// constructing an `n_vars + 1`-variate polynomial whose coefficients in the "new variable" are the
+/// individual sum claims and evaluating it at a random point. Due to linearity of sums each claim
+/// can be proven separately with an individual [`SumcheckProver`] followed by weighted summation of
+/// the round polynomials.
 pub fn batch_prove<F, Prover, Challenger_>(
 	mut provers: Vec<Prover>,
 	transcript: &mut ProverTranscript<Challenger_>,
