+++ crates/verifier/src/protocols/sumcheck/mod.rs
@@ -1,9 +1,11 @@
 // Copyright 2023-2025 Irreducible Inc.
 
+mod batch;
 pub mod common;
 mod error;
 mod verify;
 
+pub use batch::*;
 pub use common::{RoundCoeffs, RoundProof};
 pub use error::*;
 pub use verify::*;
